refactor: Add common errorCode
