+++ src/main/java/mju/sw/micro/global/adapter/MailService.java
@@ -1,4 +1,4 @@
-package mju.sw.micro.global.utils;
+package mju.sw.micro.global.adapter;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -10,22 +10,24 @@ import jakarta.mail.internet.InternetAddress;
 import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import mju.sw.micro.global.common.response.ApiResponse;
+import mju.sw.micro.global.error.exception.ErrorCode;
 
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class MailUtil {
+public class MailService {
 
 	@Value("${spring.mail.username}")
 	private String senderEmail;
 	private final JavaMailSender emailSender;
 
 	public void sendMessage(String recipientEmail, String title, String content, String extra) {
-		MimeMessage message = createMessage(recipientEmail, title, content, extra);
-		emailSender.send(message);
+		ApiResponse<MimeMessage> message = createMessage(recipientEmail, title, content, extra);
+		emailSender.send(message.getData());
 	}
 
-	private MimeMessage createMessage(String recipientEmail, String title, String content,
+	private ApiResponse<MimeMessage> createMessage(String recipientEmail, String title, String content,
 		String extra) {
 		MimeMessage message = emailSender.createMimeMessage();
 		try {
@@ -34,10 +36,10 @@ public class MailUtil {
 			message.setSubject(title); //제목
 			message.setText(contents, "UTF-8", "html"); //내용
 			message.setFrom(new InternetAddress(senderEmail)); //보내는 사람
-			return message;
+			return ApiResponse.ok(message);
 		} catch (MessagingException e) {
 			log.info("fail");
-			throw new RuntimeException(e);
+			return ApiResponse.withError(ErrorCode.BAD_REQUEST);
 		}
 	}
