+++ src/main/java/mju/sw/micro/global/error/exception/ErrorCode.java
@@ -11,6 +11,11 @@ public enum ErrorCode {
 
 	// Common
 	INVALID_INPUT_VALUE(HttpStatus.BAD_REQUEST, "Request Body를 통해 전달된 값이 유효하지 않습니다."),
+	BAD_REQUEST(HttpStatus.BAD_REQUEST, "잘못된 요청입니다."),
+	INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "서버 내부 에러가 발생했습니다"),
+	NOT_FOUND(HttpStatus.NOT_FOUND, "요청한 리소스를 찾을 수 없습니다."),
+	FORBIDDEN(HttpStatus.FORBIDDEN, "권한이 없는 접근입니다."),
+	UNAUTHORIZED(HttpStatus.UNAUTHORIZED, "인증이 필요한 접근입니다."),
 
 	// Club
 	INVALID_CLUB_ID(HttpStatus.BAD_REQUEST, "Request Body를 통해 전달된 학생 단체의 식별자가 유효하지 않습니다."),
